forgot to turn -Wall back on

+++ balloons/api.c
@@ -64,7 +64,7 @@ void exec_commands(events *e, damn *d, packet *p) {
     char *trigger = setting_get(BKEY_TRIGGER);
     if (strncmp(trigger, sp->body, strlen(trigger)) == 0) {
         char *bod = sp->body + strlen(trigger);
-        size_t len = -1;
+        size_t len = (unsigned long)-1;
         while(bod[++len] > 32);
         char *cmdname = calloc(1, len + 5);
         snprintf(cmdname, len + 5, "cmd.%s", bod);
